+++ p2p/pex/pex_reactor_test.go
@@ -109,7 +109,7 @@ func TestPEXReactorRunning(t *testing.T) {
 	addOtherNodeAddrToAddrBook(2, 1)
 
 	for i, sw := range switches {
-		sw.AddListener(p2p.NewDefaultListener("tcp", sw.NodeInfo().ListenAddr, false, logger.With("pex", i)))
+		sw.AddListener(p2p.NewDefaultListener("tcp://"+sw.NodeInfo().ListenAddr, "", false, logger.With("pex", i)))
 
 		err := sw.Start() // start switch and reactors
 		require.Nil(t, err)
@@ -229,12 +229,7 @@ func TestPEXReactorUsesSeedsIfNeeded(t *testing.T) {
 		},
 	)
 	seed.AddListener(
-		p2p.NewDefaultListener(
-			"tcp",
-			seed.NodeInfo().ListenAddr,
-			false,
-			log.TestingLogger(),
-		),
+		p2p.NewDefaultListener("tcp://"+seed.NodeInfo().ListenAddr, "", false, log.TestingLogger()),
 	)
 	require.Nil(t, seed.Start())
 	defer seed.Stop()